Support button inputs on linux host
This commit adds polling for user input when running the engine on host.
It's currently unpolished but functional. It works by polling for
keystrokes and mapping those to the Kywy hardware buttons. The WASD keys
are mapped to the D-Pad and Q and E are mapped to the left and right
buttons respectively. We also update the host build to use the new 128 x
128 serial display driver so it will mirror the size of the Kywy screen.